ref: Remove EMS\CommonBundle\Common\Standard\Base64
