+++ EMS/common-bundle/src/Twig/CommonExtension.php
@@ -4,8 +4,8 @@ namespace EMS\CommonBundle\Twig;
 
 use EMS\CommonBundle\Common\Converter;
 use EMS\CommonBundle\Common\EMSLink;
-use EMS\CommonBundle\Common\Standard\Base64;
 use EMS\CommonBundle\Helper\Text\Encoder;
+use EMS\Helpers\Standard\Base64;
 use EMS\Helpers\Standard\Color;
 use Ramsey\Uuid\Uuid;
 use Twig\Extension\AbstractExtension;
